Change default lease name for vpa-recommender to "vpa-recommender-lease" to avoid conflicts
See kubernetes/autoscaler#7461.

The default lock name can sometimes conflict with a managed deployment of VPA/HPA which can cause issues on your cluster.

Here we are proposing changing the flag value in the Helm chart to a new name (`vpa-recommender-lease`). We are keeping the old resource name in the RBAC to avoid disruptions or issues during upgrade.